Update all pages and share the code
